requirements.txt
# Core RAG pipeline
llama-index-core==0.14.15
llama-index-llms-openai==0.6.23
llama-index-embeddings-openai==0.5.1
llama-index-vector-stores-chroma==0.5.5

# Vector store
chromadb==1.5.2

# Re-ranker
sentence-transformers==5.3.0
torch==2.11.0

# LLM client
openai==2.24.0

# Environment variables
python-dotenv==1.2.1

# API server
fastapi==0.128.8
uvicorn==0.40.0
